+++ android/src/main/kotlin/com/adapty/flutter/AdaptyFlutterPlugin.kt
@@ -6,6 +6,11 @@ import com.adapty.Adapty
 import com.adapty.api.AttributionType
 import com.adapty.api.entity.containers.DataContainer
 import com.adapty.api.entity.containers.Product
+import com.adapty.flutter.extensions.safeLet
+import com.adapty.flutter.models.AdaptyProduct
+import com.adapty.flutter.models.GetActivePurchasesResult
+import com.adapty.flutter.models.GetPaywallsResult
+import com.adapty.flutter.models.MakePurchaseResult
 import com.google.gson.Gson
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
@@ -15,11 +20,6 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import com.adapty.flutter.extensions.safeLet
-import com.adapty.flutter.models.AdaptyProduct
-import com.adapty.flutter.models.GetActivePurchasesResult
-import com.adapty.flutter.models.GetPaywallsResult
-import com.adapty.flutter.models.MakePurchaseResult
 import org.jetbrains.annotations.NotNull
 
 class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
@@ -31,19 +31,19 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
     private lateinit var channel: MethodChannel
 
     private var activity: Activity? = null
-    private var methods = HashSet<String>()
+    private var results = HashSet<Int>()
 
     private var containers = ArrayList<DataContainer>()
     private var products = HashMap<String, Product>()
 
     override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, CHANNEL_NAME)
-        channel.setMethodCallHandler(this);
+        channel.setMethodCallHandler(this)
     }
 
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
-        if (!methods.contains(call.method)) {
-            methods.add(call.method)
+        if (!results.contains(result.hashCode())) {
+            results.add(result.hashCode())
             when (call.method) {
                 "activate" -> handleActivate(call, result)
                 "get_paywalls" -> handleGetPaywalls(call, result)
@@ -53,6 +53,7 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
                 "get_purchaser_info" -> handleGetPurchaserInfo(call, result)
                 "get_active_purchases" -> handleGetActivePurchases(call, result)
                 "update_attribution" -> handleUpdateAttribution(call, result)
+                "make_deferred_purchase" -> resultIfNeeded(result) { result.error(call.method, "Not implemented", null) }
                 else -> result.notImplemented()
             }
         }
@@ -88,53 +89,53 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
         val appKey: String = call.argument<String>("app_key") ?: ""
         val customerUserId: String = call.argument<String>("customer_user_id") ?: ""
         when {
-            appKey.isBlank() -> resultIfNeeded(call) { result.success(false) }
+            appKey.isBlank() -> resultIfNeeded(result) { result.success(false) }
             customerUserId.isBlank() -> {
                 activity?.let {
                     Adapty.activate(it.applicationContext, appKey)
-                    resultIfNeeded(call) { result.success(true) }
-                } ?: resultIfNeeded(call) { result.success(false) }
+                    resultIfNeeded(result) { result.success(true) }
+                } ?: resultIfNeeded(result) { result.success(false) }
             }
             else -> {
                 activity?.let {
                     Adapty.activate(it.applicationContext, appKey, customerUserId)
-                    resultIfNeeded(call) { result.success(true) }
-                } ?: resultIfNeeded(call) { result.success(false) }
+                    resultIfNeeded(result) { result.success(true) }
+                } ?: resultIfNeeded(result) { result.success(false) }
             }
         }
     }
 
     private fun handleGetPaywalls(@NonNull call: MethodCall, @NonNull result: Result) {
         activity?.let { activity ->
-            Adapty.getPurchaseContainers(activity) { containers: ArrayList<DataContainer>, products: ArrayList<Product>, state: String, error: String? ->
+            Adapty.getPurchaseContainers(activity) { containers: ArrayList<DataContainer>, products: ArrayList<Product>, _: String, error: String? ->
                 try {
                     error?.let {
-                        resultIfNeeded(call) { result.error(call.method, it, null) }
+                        resultIfNeeded(result) { result.error(call.method, it, null) }
                     } ?: run {
                         cachePaywalls(containers)
                         cacheProducts(products)
 
                         val getPaywallsResultJson = Gson().toJson(GetPaywallsResult(paywallsIds(containers), products(products)))
-                        resultIfNeeded(call) { result.success(getPaywallsResultJson) }
+                        resultIfNeeded(result) { result.success(getPaywallsResultJson) }
                     }
                 } catch (fe: FlutterException) {
-                    resultIfNeeded(call) { result.error(call.method, fe.message, fe.localizedMessage) }
+                    resultIfNeeded(result) { result.error(call.method, fe.message, fe.localizedMessage) }
                 }
             }
-        } ?: resultIfNeeded(call) { result.success("") }
+        } ?: resultIfNeeded(result) { result.success("") }
     }
 
     private fun handleMakePurchase(@NonNull call: MethodCall, @NonNull result: Result) {
         val productId: String = call.argument<String>("product_id") ?: ""
         safeLet(activity, products[productId]) { activity, product ->
-            Adapty.makePurchase(activity, product) { purchase, response, error ->
+            Adapty.makePurchase(activity, product) { purchase, _, error ->
                 error?.let {
-                    resultIfNeeded(call) { result.error(call.method, it, null) }
+                    resultIfNeeded(result) { result.error(call.method, it, null) }
                 } ?: run {
-                    resultIfNeeded(call) { result.success(Gson().toJson(MakePurchaseResult(purchase?.purchaseToken, product.skuDetails?.type))) }
+                    resultIfNeeded(result) { result.success(Gson().toJson(MakePurchaseResult(purchase?.purchaseToken, product.skuDetails?.type))) }
                 }
             }
-        } ?: resultIfNeeded(call) { result.error(call.method, "No product id passed", null) }
+        } ?: resultIfNeeded(result) { result.error(call.method, "No product id passed", null) }
     }
 
     private fun handleValidatePurchase(@NonNull call: MethodCall, @NonNull result: Result) {
@@ -143,13 +144,13 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
         val purchaseToken: String = call.argument<String>("purchase_token") ?: ""
 
         if (purchaseType.isNotBlank() && productId.isNotBlank() && purchaseToken.isNotBlank()) {
-            Adapty.validatePurchase(purchaseType, productId, purchaseToken) { response, error ->
+            Adapty.validatePurchase(purchaseType, productId, purchaseToken) { _, error ->
                 error?.let {
-                    resultIfNeeded(call) { result.error(call.method, it, null) }
-                } ?: resultIfNeeded(call) { result.success(true) }
+                    resultIfNeeded(result) { result.error(call.method, it, null) }
+                } ?: resultIfNeeded(result) { result.success(true) }
             }
         } else {
-            resultIfNeeded(call) { result.success(false) }
+            resultIfNeeded(result) { result.success(false) }
         }
     }
 
@@ -157,8 +158,8 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
         activity?.let { activity ->
             Adapty.restorePurchases(activity) { _, error ->
                 error?.let {
-                    resultIfNeeded(call) { result.error(call.method, it, null) }
-                } ?: resultIfNeeded(call) { result.success(true) }
+                    resultIfNeeded(result) { result.error(call.method, it, null) }
+                } ?: resultIfNeeded(result) { result.success(true) }
             }
         }
     }
@@ -167,8 +168,8 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
     private fun handleGetPurchaserInfo(@NotNull call: MethodCall, @NotNull result: Result) {
         Adapty.getPurchaserInfo { _, _, error ->
             error?.let {
-                resultIfNeeded(call) { result.error(call.method, it, null) }
-            } ?: resultIfNeeded(call) {
+                resultIfNeeded(result) { result.error(call.method, it, null) }
+            } ?: resultIfNeeded(result) {
                 result.success(true)
             }
         }
@@ -177,7 +178,7 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
     private fun handleGetActivePurchases(@NotNull call: MethodCall, @NotNull result: Result) {
         Adapty.getPurchaserInfo { purchaserInfo, _, error ->
             error?.let {
-                resultIfNeeded(call) { result.error(call.method, it, null) }
+                resultIfNeeded(result) { result.error(call.method, it, null) }
             } ?: run {
                 val nonSubscriptionsIds = ArrayList<String>()
                 purchaserInfo?.nonSubscriptions?.values?.forEach { nonSubscriptions ->
@@ -189,15 +190,15 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
                 val paidAccessLevel: String = call.argument<String>("paid_access_level") ?: ""
                 if (paidAccessLevel.isNotBlank()) {
                     purchaserInfo?.paidAccessLevels?.get(paidAccessLevel)?.let { subscription ->
-                        resultIfNeeded(call) {
+                        resultIfNeeded(result) {
                             result.success(Gson().toJson(GetActivePurchasesResult(subscription.isActive
                                     ?: false, subscription.vendorProductId, nonSubscriptionsIds)))
                         }
-                    } ?: resultIfNeeded(call) {
+                    } ?: resultIfNeeded(result) {
                         result.success(Gson().toJson(GetActivePurchasesResult(false, null, nonSubscriptionsIds)))
                     }
                 } else {
-                    resultIfNeeded(call) { result.success(Gson().toJson(GetActivePurchasesResult(false, null, nonSubscriptionsIds))) }
+                    resultIfNeeded(result) { result.success(Gson().toJson(GetActivePurchasesResult(false, null, nonSubscriptionsIds))) }
                 }
             }
         }
@@ -218,8 +219,8 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
                 Adapty.updateAttribution(attr, source, id)
             } ?: Adapty.updateAttribution(attr, source)
 
-            resultIfNeeded(call) { result.success(true) }
-        } ?: resultIfNeeded(call) {
+            resultIfNeeded(result) { result.success(true) }
+        } ?: resultIfNeeded(result) {
             result.success(false)
         }
     }
@@ -261,9 +262,9 @@ class AdaptyFlutterPlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
         AdaptyProduct(id, title, description, price, localizedPrice, currencyCode)
     }
 
-    private inline fun resultIfNeeded(@NonNull call: MethodCall, inline: () -> Unit) {
-        if (methods.contains(call.method)) {
-            methods.remove(call.method)
+    private inline fun resultIfNeeded(@NotNull result: Result, inline: () -> Unit) {
+        if (results.contains(result.hashCode())) {
+            results.remove(result.hashCode())
             inline.invoke()
         }
     }