+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.2
+### Fix android reply already submitted error
+### Added deferred purchases support
+
 ## 0.0.1
 
 ### Added